code/functionapps/modelDeployment/requirements.txt
azure-functions
joblib
numpy
azureml-defaults
scikit-learn